requirements.txt
Django
djangorestframework
django-ckeditor
googletrans==4.0.0-rc1
